ai-services/requirements.txt
flask==2.3.2
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3
scipy==1.10.1
nltk==3.8.1
spacy==3.6.1
networkx==3.1
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0
requests==2.31.0
python-dotenv==1.0.0
pymongo==4.4.1
redis==4.6.0
celery==5.3.1
gunicorn==21.2.0
flask-cors==4.0.0
flask-limiter==3.3.1
python-jose==3.3.0
passlib==1.7.4

# Para processamento de linguagem natural
textblob==0.17.1
wordcloud==1.9.2

# Para análise de grafos
python-igraph==0.10.6

# Para machine learning avançado
transformers==4.32.1
torch==2.0.1

# Para desenvolvimento
pytest==7.4.0
black==23.7.0
flake8==6.0.0
